scrappers/requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
future==0.17.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.1
mysql-connector-python==8.0.17
parsel==1.5.2
Pillow==6.1.0
protobuf==3.9.1
proxyscrape==0.3.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyInstaller==3.5
pyOpenSSL==19.0.0
queuelib==1.5.0
requests==2.22.0
Scrapy==1.7.3
scrapy-inline-requests==0.3.1
scrapy-proxy-pool==0.1.6
scrapy-rotated-proxy==0.1.5
Scrapy-UserAgents==0.0.1
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.3
Twisted==19.7.0
urllib3==1.25.3
w3lib==1.21.0
zope.interface==4.6.0
